Move utility function to read streams to ResourceProvider
