+++ src/main/java/edu/kit/provideq/toolbox/ResourceProvider.java
@@ -1,13 +1,17 @@
 package edu.kit.provideq.toolbox;
 
 import edu.kit.provideq.toolbox.meta.ProblemType;
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.io.Resource;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.stereotype.Component;
 
@@ -61,6 +65,58 @@ public class ResourceProvider {
   }
 
   /**
+   * Reads the input stream of a {@link BufferedReader} into a string.
+   *
+   * @param reader reader to read from
+   * @return full string of the input stream
+   * @throws IOException when the input stream couldn't be read
+   */
+  public String readStream(BufferedReader reader) throws IOException {
+    var inputBuilder = new StringBuilder();
+
+    // Process first line manually to avoid adding a newline at the beginning
+    var line = reader.readLine();
+    if (line != null) {
+      inputBuilder.append(line);
+    }
+
+    line = reader.readLine();
+    while (line != null) {
+      inputBuilder.append('\n').append(line);
+      line = reader.readLine();
+    }
+    reader.close();
+
+    return inputBuilder.toString();
+  }
+
+  /**
+   * Reads the input stream of an {@link InputStream} into a string.
+   *
+   * @param stream stream to read from
+   * @return full string of the input stream
+   * @throws IOException when the input stream couldn't be read
+   */
+  public String readStream(InputStream stream) throws IOException {
+    return readStream(new BufferedReader(new InputStreamReader(stream)));
+  }
+
+  /**
+   * Reads the resource at the specified resource path into a string.
+   *
+   * @param resourcePath path that points to the requested resource
+   * @return full string of the data inside the resource
+   * @throws IOException when the resource is not available or couldn't be read
+   */
+  public String readResourceString(String resourcePath) throws IOException {
+    Resource resource = resourceLoader.getResource("classpath:" + resourcePath);
+
+    try (var reader = new BufferedReader(new InputStreamReader(resource.getInputStream()))) {
+      return readStream(reader);
+    }
+  }
+
+  /**
    * Returns the resource at the specified resource path.
    *
    * @param resourcePath path that points to the requested resource
@@ -68,11 +124,8 @@ public class ResourceProvider {
    * @throws IOException when the resource is not available
    */
   public File getResource(String resourcePath) throws IOException {
-    return getRootFile(resourcePath);
-
-    /* removed as long as we're not using resources directly
     Resource resource = resourceLoader.getResource("classpath:" + resourcePath);
-    return resource.getFile();*/
+    return resource.getFile();
   }
 
   /**
